factor out lantern specific code
